pyproject.toml
[project]
name = "helm"
version = "0.1.0"
description = "Observation and evaluation framework for multi-agent AI systems"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "httpx-sse>=0.4",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "typer>=0.12",
    "kagglehub>=0.4.3",
]

[project.scripts]
helm = "helm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/helm"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
